Move gossip structs into behaviour
